mega refactoring audio folder

+++ plugins/components/EventInterface.h
@@ -8,6 +8,6 @@ class EventInterface {
 public:
     virtual void onMotion(MotionInterface& motion) = 0;
     virtual void onMotionRelease(MotionInterface& motion) = 0;
-    virtual void onEncoder(int id, int8_t direction, uint32_t tick) = 0;
+    virtual void onEncoder(int id, int8_t direction, uint64_t tick) = 0;
     virtual void onKey(uint16_t id, int key, int8_t state) = 0;
 };
